pyproject.toml
[project]
name = "llmtranslate"
version = "1.0.0"
description = "LLM Vision APIを使ったWindows向けスクリーン翻訳アプリ"
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.6.0",
    "mss>=9.0.0",
    "Pillow>=10.0.0",
    "httpx>=0.27.0",
    "numpy>=1.24.0",
    "rapidocr-onnxruntime>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
